backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
aiofiles==23.2.1

# Database
motor==3.3.2
pymongo==4.6.0

# ML/NLP Dependencies
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
transformers==4.36.2
spacy==3.7.2
nltk==3.8.1
numpy==1.24.3
scipy==1.11.4

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Google Calendar API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Microsoft Graph API
msal==1.26.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2